Double value for grades

+++ GradeParser.BL/Data/Model/Grade.cs
@@ -2,8 +2,8 @@
 {
     public class Grade
     {
-        public int ClassicGrade { get; set; }
-        public int BolognaGrade { get; set; }
+        public double ClassicGrade { get; set; }
+        public double BolognaGrade { get; set; }
         public string ESTCGrade { get; set; }
         
     }
